Group 'Tag' and 'Card' stories into 'Content', and sort it
