+++ .storybook/preview.js
@@ -18,6 +18,29 @@ export const parameters = {
           ],
           'Toggle',
         ],
+        'Content', [
+          'Card', [
+            'Filled Card',
+            'Outline Card',
+            'Outline Filled Card',
+            'With Double Action',
+            'With Link Action',
+            'With Overflowed Image',
+            'With Overlaid Header',
+            'Card Subcomponents', [
+              'Card Divider',
+              'Card Header',
+              'Card Image',
+              'Double Action',
+              'Link Action',
+            ],
+          ],
+          'Tag', [
+            'Simple Tag',
+            'Dropdown Tag',
+            'Removable Tag',
+          ],
+        ],
       ],
     },
   },
