+++ stories/Content/Card/OutlineFilledCard.stories.js
@@ -1,13 +1,13 @@
 import React from "react";
 
 import {
-  FilledCard as Component,
+  OutlineFilledCard as Component,
   CardHeader,
   CardImage,
   CardDivider,
   DoubleAction,
-} from "../../confetti-ds/src/Card";
-import { filledCardChoices } from "../../confetti-ds/src/Card/propTypes";
+} from "../../../confetti-ds/src/Card";
+import { outlineFilledCardChoices } from "../../../confetti-ds/src/Card/propTypes";
 
 import { DoubleAction as DoubleActionStory } from "./subcomponents/DoubleAction.stories";
 import { CardDivider as CardDividerStory } from "./subcomponents/CardDivider.stories";
@@ -15,15 +15,12 @@ import { CardImage as CardImageStory } from "./subcomponents/CardImage.stories";
 import { CardHeader as CardHeaderStory } from "./subcomponents/CardHeader.stories";
 
 export default {
-  title: "Card/Filled Card",
+  title: "Content/Card/Outline Filled Card",
   component: Component,
   subcomponents: { CardImage, CardHeader, CardDivider, DoubleAction },
   argTypes: {
     color: {
-      control: { type: "radio", options: filledCardChoices.color },
-    },
-    skin: {
-      control: { type: "radio", options: filledCardChoices.skin },
+      control: { type: "radio", options: outlineFilledCardChoices.color },
     },
     children: {
       control: false,
@@ -31,7 +28,7 @@ export default {
   },
 };
 
-export const FilledCard = (args) => (
+export const OutlineFilledCard = (args) => (
   <div className="ignore-this-div" style={{ width: "400px", margin: "0 auto" }}>
     <Component {...args}>
       <CardHeaderStory {...CardHeaderStory.args} />
@@ -46,7 +43,6 @@ export const FilledCard = (args) => (
   </div>
 );
 
-FilledCard.args = {
+OutlineFilledCard.args = {
   color: "mineral",
-  skin: "pale",
-};
+}
