+++ stories/Content/Card/WithOverflowedImage.stories.js
@@ -6,8 +6,8 @@ import {
   CardImage,
   CardDivider,
   DoubleAction,
-} from "../../confetti-ds/src/Card";
-import { outlineCardChoices } from "../../confetti-ds/src/Card/propTypes";
+} from "../../../confetti-ds/src/Card";
+import { outlineCardChoices } from "../../../confetti-ds/src/Card/propTypes";
 
 import { DoubleAction as DoubleActionStory } from "./subcomponents/DoubleAction.stories";
 import { CardDivider as CardDividerStory } from "./subcomponents/CardDivider.stories";
@@ -15,7 +15,7 @@ import { CardImage as CardImageStory } from "./subcomponents/CardImage.stories";
 import { CardHeader as CardHeaderStory } from "./subcomponents/CardHeader.stories";
 
 export default {
-  title: "Card/With Overlaid Header",
+  title: "Content/Card/With Overflowed Image",
   component: Component,
   subcomponents: { CardImage, CardHeader, CardDivider, DoubleAction },
   argTypes: {
@@ -31,12 +31,11 @@ export default {
   },
 };
 
-export const WithOverlaidHeader = (args) => (
+export const WithOverflowedImage = (args) => (
   <div className="ignore-this-div" style={{ width: "400px", margin: "0 auto" }}>
-    {/* for overlaid headers, card should not be horizontal */}
-    <Component {...args} isHorizontal={false}>
+    <Component {...args}>
       <CardImageStory {...CardImageStory.args} isOverflowed={true} />
-      <CardHeaderStory {...CardHeaderStory.args} isOverlay={true} />
+      <CardHeaderStory {...CardHeaderStory.args} />
 
       <p>This HTML is inside the card body, below the image and header.</p>
       <p>We'd recommend for it to not be too long.</p>
@@ -47,7 +46,7 @@ export const WithOverlaidHeader = (args) => (
   </div>
 );
 
-WithOverlaidHeader.args = {
+WithOverflowedImage.args = {
   color: "mineral",
   skin: "pale"
 }
