+++ stories/Content/Card/WithOverlaidHeader.stories.js
@@ -5,19 +5,19 @@ import {
   CardHeader,
   CardImage,
   CardDivider,
-  LinkAction,
-} from "../../confetti-ds/src/Card";
-import { outlineCardChoices } from "../../confetti-ds/src/Card/propTypes";
+  DoubleAction,
+} from "../../../confetti-ds/src/Card";
+import { outlineCardChoices } from "../../../confetti-ds/src/Card/propTypes";
 
-import { LinkAction as LinkActionStory } from "./subcomponents/LinkAction.stories";
+import { DoubleAction as DoubleActionStory } from "./subcomponents/DoubleAction.stories";
 import { CardDivider as CardDividerStory } from "./subcomponents/CardDivider.stories";
 import { CardImage as CardImageStory } from "./subcomponents/CardImage.stories";
 import { CardHeader as CardHeaderStory } from "./subcomponents/CardHeader.stories";
 
 export default {
-  title: "Card/With Link Action",
+  title: "Content/Card/With Overlaid Header",
   component: Component,
-  subcomponents: { CardImage, CardHeader, CardDivider, LinkAction },
+  subcomponents: { CardImage, CardHeader, CardDivider, DoubleAction },
   argTypes: {
     color: {
       control: { type: "radio", options: outlineCardChoices.color },
@@ -31,23 +31,23 @@ export default {
   },
 };
 
-export const WithLinkAction = (args) => (
+export const WithOverlaidHeader = (args) => (
   <div className="ignore-this-div" style={{ width: "400px", margin: "0 auto" }}>
-    <Component {...args}>
-      <CardHeaderStory {...CardHeaderStory.args} />
+    {/* for overlaid headers, card should not be horizontal */}
+    <Component {...args} isHorizontal={false}>
+      <CardImageStory {...CardImageStory.args} isOverflowed={true} />
+      <CardHeaderStory {...CardHeaderStory.args} isOverlay={true} />
 
       <p>This HTML is inside the card body, below the image and header.</p>
       <p>We'd recommend for it to not be too long.</p>
 
-      <CardImageStory {...CardImageStory.args} />
       <CardDividerStory {...CardDividerStory.args} />
-      <LinkActionStory {...LinkActionStory.args} />
+      <DoubleAction {...DoubleActionStory.args} />
     </Component>
   </div>
 );
 
-WithLinkAction.args = {
+WithOverlaidHeader.args = {
   color: "mineral",
   skin: "pale"
 }
-
