+++ stories/Content/Card/subcomponents/CardDivider.stories.js
@@ -1,9 +1,9 @@
 import React from "react";
 
-import { CardDivider as Component } from "../../../confetti-ds/src/Card";
+import { CardDivider as Component } from "../../../../confetti-ds/src/Card";
 
 export default {
-  title: "Card/Subcomponents/Card Divider",
+  title: "Content/Card/Card Subcomponents/Card Divider",
   component: Component,
 };
