+++ stories/Content/Card/subcomponents/CardHeader.stories.js
@@ -1,10 +1,10 @@
 import React from "react";
 
-import { CardHeader as Component } from "../../../confetti-ds/src/Card";
-import { ICON_TYPES } from "../../../confetti-ds/src/constants";
+import { CardHeader as Component } from "../../../../confetti-ds/src/Card";
+import { ICON_TYPES } from "../../../../confetti-ds/src/constants";
 
 export default {
-  title: "Card/Subcomponents/Card Header",
+  title: "Content/Card/Card Subcomponents/Card Header",
   component: Component,
   argTypes: {
     categoryIcon: {
