+++ stories/Content/Card/subcomponents/card-image.jpg
