+++ stories/Content/Tag/DropdownTag/DropdownTag.stories.js
@@ -1,10 +1,10 @@
 import React from "react";
 
-import { TAG_COLORS } from "../../confetti-ds/src/constants";
-import { DropdownTag as Component  } from "../../confetti-ds/src";
+import { TAG_COLORS } from "../../../../confetti-ds/src/constants";
+import { DropdownTag as Component  } from "../../../../confetti-ds/src";
 
 export default {
-  title: "Tag/Dropdown Tag",
+  title: "Content/Tag/Dropdown Tag/Dropdown Tag",
   component: Component,
   argTypes: {
     color: {
