+++ stories/Content/Tag/TogglableTag.stories.js
@@ -1,10 +1,10 @@
 import React from "react";
 
-import { TAG_COLORS } from "../../confetti-ds/src/constants";
-import { TogglableTag as Component } from "../../confetti-ds/src";
+import { TAG_COLORS } from "../../../confetti-ds/src/constants";
+import { TogglableTag as Component } from "../../../confetti-ds/src";
 
 export default {
-  title: "Tag/Togglable Tag",
+  title: "Content/Tag/Togglable Tag",
   component: Component,
   argTypes: {
     color: {
